+++ CHANGELOG.md
@@ -22,6 +22,15 @@ TODO: Date
     struct DyncastWrapper<T>(pub T);
     ```
 
+  - It's now possible to combine a reinterpret-cast and unsizing:
+
+     ```rust
+     #[derive(Dyncast)]
+     #[dyncast(unsafe T as dyn Trait)]
+     #[repr(transparent)]
+     struct DyncastWrapper<T: Trait>(pub T);
+     ```
+
 - Revisions:
   - Fixed issue where a bunch of code would be printed into the documentation instead of just the type.
   - Dyncast targets emitted into the documentation are now links.